HACK: scale bin wall to 1e-08 thiccness so double rounds to 0 as float in vertex
Vertices are Vector3[] which is an array of floats.
A bin wall is an 3D object with 8 vertices, only want 2D with 4 vertices.
By scaling bin wall to 1e-08 thiccness so double rounds to 0 as float in vertex,
then the 3D bin wall becomes a 2D bin wall in vertices (duplicate of 4x2 stacked)
This way when combining mesh vertices there will be one common local point for the agent to pick from.
Reducing complexity.